Add optional --device-name argument
This allows specifying the name of the USB to write to in advance, which
means the program can be used non-interactively.

The device name is quite specific to the device (it contains the
manufacter, product, and serial, e.g. "Samsung PSSD T7 S1SLVX2T1210"),
so it doesn't carry the same risk as using a device path such as
"/dev/sdc" (which you could rather easily typo and wipe out an internal
system disk).